Fix NonBrowser migration
