+++ src/migration.rs
@@ -610,15 +610,15 @@ where
 fn migrate_0_41_events(store: &super::Store) -> Result<()> {
     use migration_structures::{
         CryptocurrencyMiningPoolV0_39, ExternalDdosV0_39, FtpBruteForceV0_39, FtpPlainTextV0_39,
-        LdapBruteForceV0_39, LdapPlainTextV0_39, MultiHostPortScanV0_39, NonBrowserV0_39,
-        PortScanV0_39, RdpBruteForceV0_39, RepeatedHttpSessionsV0_39,
+        LdapBruteForceV0_39, LdapPlainTextV0_39, MultiHostPortScanV0_39, PortScanV0_39,
+        RdpBruteForceV0_39, RepeatedHttpSessionsV0_39,
     };
     use num_traits::FromPrimitive;
 
     use crate::event::{
         BlocklistConnFields, CryptocurrencyMiningPool, EventKind, ExternalDdos, FtpBruteForce,
         FtpPlainText, HttpEventFieldsV0_39, HttpEventFieldsV0_41, LdapBruteForce, LdapPlainText,
-        MultiHostPortScan, NonBrowser, PortScan, RdpBruteForce, RepeatedHttpSessions,
+        MultiHostPortScan, PortScan, RdpBruteForce, RepeatedHttpSessions,
     };
 
     let event_db = store.events();
@@ -647,13 +647,15 @@ fn migrate_0_41_events(store: &super::Store) -> Result<()> {
                 let new_value = bincode::serialize(&fields).unwrap_or_default();
                 event_db.update((&k, &v), (&k, &new_value))?;
             }
-            EventKind::TorConnection => {
+            EventKind::NonBrowser => {
                 update_event_db_with_new_event::<HttpEventFieldsV0_39, HttpEventFieldsV0_41>(
                     &k, &v, &event_db,
                 )?;
             }
-            EventKind::NonBrowser => {
-                update_event_db_with_new_event::<NonBrowserV0_39, NonBrowser>(&k, &v, &event_db)?;
+            EventKind::TorConnection => {
+                update_event_db_with_new_event::<HttpEventFieldsV0_39, HttpEventFieldsV0_41>(
+                    &k, &v, &event_db,
+                )?;
             }
             EventKind::RepeatedHttpSessions => {
                 update_event_db_with_new_event::<RepeatedHttpSessionsV0_39, RepeatedHttpSessions>(
