+++ src/migration/migration_structures.rs
@@ -722,84 +722,6 @@ impl From<RepeatedHttpSessionsV0_39> for crate::event::RepeatedHttpSessions {
 }
 
 #[derive(Debug, Deserialize, Serialize)]
-pub struct NonBrowserV0_39 {
-    #[serde(with = "ts_nanoseconds")]
-    pub time: DateTime<Utc>,
-    pub sensor: String,
-    #[serde(with = "ts_nanoseconds")]
-    pub session_end_time: DateTime<Utc>,
-    pub src_addr: IpAddr,
-    pub src_port: u16,
-    pub dst_addr: IpAddr,
-    pub dst_port: u16,
-    pub proto: u8,
-    pub method: String,
-    pub host: String,
-    pub uri: String,
-    pub referer: String,
-    pub version: String,
-    pub user_agent: String,
-    pub request_len: usize,
-    pub response_len: usize,
-    pub status_code: u16,
-    pub status_msg: String,
-    pub username: String,
-    pub password: String,
-    pub cookie: String,
-    pub content_encoding: String,
-    pub content_type: String,
-    pub cache_control: String,
-    pub orig_filenames: Vec<String>,
-    pub orig_mime_types: Vec<String>,
-    pub resp_filenames: Vec<String>,
-    pub resp_mime_types: Vec<String>,
-    pub post_body: Vec<u8>,
-    pub state: String,
-    pub category: EventCategory,
-    pub triage_scores: Option<Vec<TriageScore>>,
-}
-
-impl From<NonBrowserV0_39> for crate::event::NonBrowser {
-    fn from(old: NonBrowserV0_39) -> Self {
-        Self {
-            time: old.time,
-            sensor: old.sensor,
-            end_time: old.session_end_time,
-            src_addr: old.src_addr,
-            src_port: old.src_port,
-            dst_addr: old.dst_addr,
-            dst_port: old.dst_port,
-            proto: old.proto,
-            method: old.method,
-            host: old.host,
-            uri: old.uri,
-            referer: old.referer,
-            version: old.version,
-            user_agent: old.user_agent,
-            request_len: old.request_len,
-            response_len: old.response_len,
-            status_code: old.status_code,
-            status_msg: old.status_msg,
-            username: old.username,
-            password: old.password,
-            cookie: old.cookie,
-            content_encoding: old.content_encoding,
-            content_type: old.content_type,
-            cache_control: old.cache_control,
-            orig_filenames: old.orig_filenames,
-            orig_mime_types: old.orig_mime_types,
-            resp_filenames: old.resp_filenames,
-            resp_mime_types: old.resp_mime_types,
-            post_body: old.post_body,
-            state: old.state,
-            confidence: 1.0,
-            category: old.category,
-            triage_scores: old.triage_scores,
-        }
-    }
-}
-
-#[derive(Debug, Deserialize, Serialize)]
 pub struct PortScanV0_39 {
     #[serde(with = "ts_nanoseconds")]
     pub time: DateTime<Utc>,
